requirements.txt
py-trello
pandas
matplotlib
pandas
xlrd
openpyxl
termcolor